# Improve performance of GET requests