--- a/msgraph/fasterclient.go
+++ b/msgraph/fasterclient.go
@@ -10,7 +10,122 @@ import (
 	"strings"
 
 	"github.com/hashicorp/go-azure-sdk/sdk/odata"
+	"github.com/hashicorp/go-retryablehttp"
 )
+
+// fasterPerformRequest is used by the package to send an HTTP request to the API.
+func (c Client) fasterPerformRequest(req *http.Request, input HttpRequestInput) (*http.Response, int, *odata.OData, error) {
+	var status int
+
+	query := input.GetOData()
+	req.Header = query.AppendHeaders(req.Header)
+	req.Header.Add("Content-Type", input.GetContentType())
+
+	if c.Authorizer != nil {
+		token, err := c.Authorizer.Token(req.Context(), req)
+		if err != nil {
+			return nil, status, nil, err
+		}
+		token.SetAuthHeader(req)
+	}
+
+	if c.UserAgent != "" {
+		req.Header.Add("User-Agent", c.UserAgent)
+	}
+
+	var resp *http.Response
+	var o *odata.OData
+	var err error
+
+	var reqBody []byte
+	if req.Body != nil {
+		reqBody, err = io.ReadAll(req.Body)
+		if err != nil {
+			return nil, status, nil, fmt.Errorf("reading request body: %v", err)
+		}
+	}
+
+	c.RetryableClient.CheckRetry = func(ctx context.Context, resp *http.Response, err error) (bool, error) {
+		if resp != nil && !c.DisableRetries {
+			if resp.StatusCode == http.StatusFailedDependency {
+				return true, nil
+			}
+
+			o, err = odata.FromResponse(resp)
+			if err != nil {
+				return false, err
+			}
+
+			f := input.GetConsistencyFailureFunc()
+			if f != nil && f(resp, o) {
+				return true, nil
+			}
+		}
+		return retryablehttp.DefaultRetryPolicy(ctx, resp, err)
+	}
+
+	req.Body = io.NopCloser(bytes.NewBuffer(reqBody))
+
+	if c.RequestMiddlewares != nil {
+		for _, m := range *c.RequestMiddlewares {
+			r, err := m(req)
+			if err != nil {
+				return nil, status, nil, err
+			}
+			req = r
+		}
+	}
+
+	resp, err = c.HttpClient.Do(req)
+	if err != nil {
+		return nil, status, nil, err
+	}
+
+	if c.ResponseMiddlewares != nil {
+		for _, m := range *c.ResponseMiddlewares {
+			r, err := m(req, resp)
+			if err != nil {
+				return nil, status, nil, err
+			}
+			resp = r
+		}
+	}
+
+	o, err = odata.FromResponse(resp)
+	if err != nil {
+		return nil, status, o, err
+	}
+	if resp == nil {
+		return resp, status, o, fmt.Errorf("nil response received")
+	}
+
+	status = resp.StatusCode
+	if !containsStatusCode(input.GetValidStatusCodes(), status) {
+		f := input.GetValidStatusFunc()
+		if f != nil && f(resp, o) {
+			return resp, status, o, nil
+		}
+
+		var errText string
+		switch {
+		case o != nil && o.Error != nil && o.Error.String() != "":
+			errText = fmt.Sprintf("OData error: %s", o.Error)
+		default:
+			defer resp.Body.Close()
+			respBody, err := io.ReadAll(resp.Body)
+			if err != nil {
+				return nil, status, o, fmt.Errorf("unexpected status %d, could not read response body", status)
+			}
+			if len(respBody) == 0 {
+				return nil, status, o, fmt.Errorf("unexpected status %d received with no body", status)
+			}
+			errText = fmt.Sprintf("response: %s", respBody)
+		}
+		return nil, status, o, fmt.Errorf("unexpected status %d with %s", status, errText)
+	}
+
+	return resp, status, o, nil
+}
 
 // FasterGet performs a GET request.
 func (c Client) FasterGet(ctx context.Context, input GetHttpRequestInput) (*http.Response, int, *odata.OData, error) {
@@ -36,7 +151,7 @@ func (c Client) FasterGet(ctx context.Context, input GetHttpRequestInput) (*http
 	}
 
 	// Perform the request
-	resp, status, o, err := c.performRequest(req, input)
+	resp, status, o, err := c.fasterPerformRequest(req, input)
 	if err != nil {
 		return nil, status, o, err
 	}