@moio 2024-02-23
@manicminer do you have any impressions to share about this work?