@manicminer 2024-09-24
Hi @moio, my apologies for the delay in responding. I have unfortunately not had the bandwidth to take a proper look at these proposed improvements. I am not sure if you are still using this project at Rancher, but regardless I would encourage you to take a look at the new [Microsoft Graph SDK published by HashiCorp](https://github.com/hashicorp/go-azure-sdk/tree/main/microsoft-graph). Whilst that project is primarily intended for use in HashiCorp projects, I am sure that performance improvements would be welcomed there, wherever compatibility is maintained. Although you may find that implementing your own API client is the better way to go - and if this is the case you might find that project useful as a reference implementation, or to use some of the base packages for authentication etc.

I am about to deprecate/archive this project in favor of the HashiCorp MS Graph SDK, and I won't be merging any more PRs at this time. I'm therefore going to have to close this, again my apologies for the delay in responding and in hindsight it would have been better to let you know about this trajectory much earlier, I am sorry about that. Thank you again for spending the time to work on this :pray: