@moio 2024-09-25
Thanks anyway.

FTR in case somebody else in future stumbles upon this, Rancher ended up keeping a forked version of this repo with this PR in as a short-term solution and migrating over to the Microsoft SDK as a long-term one.

👋